+++ crates/egui/src/widgets/button.rs
@@ -21,6 +21,7 @@ use crate::*;
 #[must_use = "You should put this widget in an ui with `ui.add(widget);`"]
 pub struct Button {
     text: WidgetText,
+    shortcut_text: WidgetText,
     wrap: Option<bool>,
     /// None means default for interact
     fill: Option<Color32>,
@@ -36,6 +37,7 @@ impl Button {
     pub fn new(text: impl Into<WidgetText>) -> Self {
         Self {
             text: text.into(),
+            shortcut_text: Default::default(),
             wrap: None,
             fill: None,
             stroke: None,
@@ -47,23 +49,16 @@ impl Button {
         }
     }
 
-    /// Creates a button with an image to the left of the text. The size of the image as displayed is defined by the size Vec2 provided.
+    /// Creates a button with an image to the left of the text. The size of the image as displayed is defined by the provided size.
     #[allow(clippy::needless_pass_by_value)]
     pub fn image_and_text(
         texture_id: TextureId,
-        size: impl Into<Vec2>,
+        image_size: impl Into<Vec2>,
         text: impl Into<WidgetText>,
     ) -> Self {
         Self {
-            text: text.into(),
-            fill: None,
-            stroke: None,
-            sense: Sense::click(),
-            small: false,
-            frame: None,
-            wrap: None,
-            min_size: Vec2::ZERO,
-            image: Some(widgets::Image::new(texture_id, size)),
+            image: Some(widgets::Image::new(texture_id, image_size)),
+            ..Self::new(text)
         }
     }
 
@@ -116,8 +111,19 @@ impl Button {
         self
     }
 
+    /// Set the minimum size of the button.
     pub fn min_size(mut self, min_size: Vec2) -> Self {
         self.min_size = min_size;
+        self
+    }
+
+    /// Show some text on the right side of the button, in weak color.
+    ///
+    /// Designed for menu buttons, for setting a keyboard shortcut text (e.g. `Ctrl+S`).
+    ///
+    /// The text can be created with [`Context::format_shortcut`].
+    pub fn shortcut_text(mut self, shortcut_text: impl Into<WidgetText>) -> Self {
+        self.shortcut_text = shortcut_text.into();
         self
     }
 }
@@ -126,6 +132,7 @@ impl Widget for Button {
     fn ui(self, ui: &mut Ui) -> Response {
         let Button {
             text,
+            shortcut_text,
             wrap,
             fill,
             stroke,
@@ -142,38 +149,39 @@ impl Widget for Button {
         if small {
             button_padding.y = 0.0;
         }
-        let total_extra = button_padding + button_padding;
 
-        let wrap_width = ui.available_width() - total_extra.x;
-        let text = text.into_galley(ui, wrap, wrap_width, TextStyle::Button);
+        let mut text_wrap_width = ui.available_width() - 2.0 * button_padding.x;
+        if let Some(image) = image {
+            text_wrap_width -= image.size().x + ui.spacing().icon_spacing;
+        }
+        if !shortcut_text.is_empty() {
+            text_wrap_width -= 60.0; // Some space for the shortcut text (which we never wrap).
+        }
 
-        let mut desired_size = text.size() + 2.0 * button_padding;
+        let text = text.into_galley(ui, wrap, text_wrap_width, TextStyle::Button);
+        let shortcut_text = (!shortcut_text.is_empty())
+            .then(|| shortcut_text.into_galley(ui, Some(false), f32::INFINITY, TextStyle::Button));
+
+        let mut desired_size = text.size();
+        if let Some(image) = image {
+            desired_size.x += image.size().x + ui.spacing().icon_spacing;
+            desired_size.y = desired_size.y.max(image.size().y);
+        }
+        if let Some(shortcut_text) = &shortcut_text {
+            desired_size.x += ui.spacing().item_spacing.x + shortcut_text.size().x;
+            desired_size.y = desired_size.y.max(shortcut_text.size().y);
+        }
         if !small {
             desired_size.y = desired_size.y.at_least(ui.spacing().interact_size.y);
         }
+        desired_size += 2.0 * button_padding;
         desired_size = desired_size.at_least(min_size);
-
-        if let Some(image) = image {
-            desired_size.x += image.size().x + ui.spacing().icon_spacing;
-            desired_size.y = desired_size.y.max(image.size().y + 2.0 * button_padding.y);
-        }
 
         let (rect, response) = ui.allocate_at_least(desired_size, sense);
         response.widget_info(|| WidgetInfo::labeled(WidgetType::Button, text.text()));
 
         if ui.is_rect_visible(rect) {
             let visuals = ui.style().interact(&response);
-            let text_pos = if let Some(image) = image {
-                let icon_spacing = ui.spacing().icon_spacing;
-                pos2(
-                    rect.min.x + button_padding.x + image.size().x + icon_spacing,
-                    rect.center().y - 0.5 * text.size().y,
-                )
-            } else {
-                ui.layout()
-                    .align_size_within_rect(text.size(), rect.shrink2(button_padding))
-                    .min
-            };
 
             if frame {
                 let fill = fill.unwrap_or(visuals.bg_fill);
@@ -186,15 +194,38 @@ impl Widget for Button {
                 );
             }
 
+            let text_pos = if let Some(image) = image {
+                let icon_spacing = ui.spacing().icon_spacing;
+                pos2(
+                    rect.min.x + button_padding.x + image.size().x + icon_spacing,
+                    rect.center().y - 0.5 * text.size().y,
+                )
+            } else {
+                ui.layout()
+                    .align_size_within_rect(text.size(), rect.shrink2(button_padding))
+                    .min
+            };
             text.paint_with_visuals(ui.painter(), text_pos, visuals);
-        }
 
-        if let Some(image) = image {
-            let image_rect = Rect::from_min_size(
-                pos2(rect.min.x, rect.center().y - 0.5 - (image.size().y / 2.0)),
-                image.size(),
-            );
-            image.paint_at(ui, image_rect);
+            if let Some(shortcut_text) = shortcut_text {
+                let shortcut_text_pos = pos2(
+                    rect.max.x - button_padding.x - shortcut_text.size().x,
+                    rect.center().y - 0.5 * shortcut_text.size().y,
+                );
+                shortcut_text.paint_with_fallback_color(
+                    ui.painter(),
+                    shortcut_text_pos,
+                    ui.visuals().weak_text_color(),
+                );
+            }
+
+            if let Some(image) = image {
+                let image_rect = Rect::from_min_size(
+                    pos2(rect.min.x, rect.center().y - 0.5 - (image.size().y / 2.0)),
+                    image.size(),
+                );
+                image.paint_at(ui, image_rect);
+            }
         }
 
         response